:ambulance: Fix SoftAP mode.
For WPA (all variants) the minimum password length is eight characters, the default password being used previously was seven characters. With this change the SoftAP name and password will now be "esp32olcbio".

Added new behavior on the User button during startup. When only the User button is held during startup the WiFi debug diagnostics will be enabled and printed to the serial console.